Refactor into separate dimensional and unit types
